# improve performance of `DGMulti` DGSEM