On a crappy notebook, I get
```julia
julia> trixi_include("examples/dgmulti_2d/elixir_euler_rayleigh_taylor_instability.jl",
           tspan=(0.0, 0.5), approximation_type=SBP())
[...]
────────────────────────────────────────────────────────────────────────────────────────────────────
Trixi simulation finished.  Final time: 0.5  Time steps: 149 (accepted), 152 (total)
────────────────────────────────────────────────────────────────────────────────────────────────────

 ───────────────────────────────────────────────────────────────────────────────────
              Trixi.jl                      Time                   Allocations      
                                    ──────────────────────   ───────────────────────
          Tot / % measured:              3.33s / 93.6%           8.60MiB / 69.1%    

 Section                    ncalls     time   %tot     avg     alloc   %tot      avg
 ───────────────────────────────────────────────────────────────────────────────────
 rhs!                        1.37k    3.03s  97.3%  2.21ms   1.16MiB  19.5%     884B
   calc_volume_integral!     1.37k    2.41s  77.3%  1.76ms     0.00B  0.00%    0.00B
   calc_interface_flux!      1.37k    466ms  14.9%   340μs     0.00B  0.00%    0.00B
   prolong2interfaces!       1.37k   33.8ms  1.08%  24.6μs     0.00B  0.00%    0.00B
   calc_surface_integral!    1.37k   31.5ms  1.01%  23.0μs     0.00B  0.00%    0.00B
   calc_sources!             1.37k   27.7ms  0.89%  20.2μs     0.00B  0.00%    0.00B
   invert_jacobian           1.37k   21.2ms  0.68%  15.5μs     0.00B  0.00%    0.00B
   Reset du/dt               1.37k   15.3ms  0.49%  11.2μs     0.00B  0.00%    0.00B
   calc_boundary_flux!       1.37k   14.7ms  0.47%  10.8μs     0.00B  0.00%    0.00B
   ~rhs!~                    1.37k   11.3ms  0.36%  8.22μs   1.16MiB  19.5%     884B
 analyze solution                3   85.4ms  2.74%  28.5ms   4.79MiB  80.5%  1.60MiB
 ───────────────────────────────────────────────────────────────────────────────────
```

on `main` and 

```julia
julia> trixi_include("examples/dgmulti_2d/elixir_euler_rayleigh_taylor_instability.jl",
           tspan=(0.0, 0.5), approximation_type=SBP())
[...]
────────────────────────────────────────────────────────────────────────────────────────────────────
Trixi simulation finished.  Final time: 0.5  Time steps: 149 (accepted), 152 (total)
────────────────────────────────────────────────────────────────────────────────────────────────────

 ───────────────────────────────────────────────────────────────────────────────────
              Trixi.jl                      Time                   Allocations      
                                    ──────────────────────   ───────────────────────
          Tot / % measured:              2.25s / 90.1%           8.69MiB / 69.4%    

 Section                    ncalls     time   %tot     avg     alloc   %tot      avg
 ───────────────────────────────────────────────────────────────────────────────────
 rhs!                        1.37k    1.94s  95.8%  1.41ms   1.24MiB  20.6%     948B
   calc_volume_integral!     1.37k    1.31s  64.5%   953μs     0.00B  0.00%    0.00B
   calc_interface_flux!      1.37k    468ms  23.1%   341μs     0.00B  0.00%    0.00B
   calc_surface_integral!    1.37k   34.8ms  1.72%  25.4μs     0.00B  0.00%    0.00B
   prolong2interfaces!       1.37k   34.4ms  1.70%  25.1μs     0.00B  0.00%    0.00B
   calc_sources!             1.37k   29.2ms  1.44%  21.3μs     0.00B  0.00%    0.00B
   invert_jacobian           1.37k   21.5ms  1.06%  15.7μs     0.00B  0.00%    0.00B
   Reset du/dt               1.37k   16.3ms  0.81%  11.9μs     0.00B  0.00%    0.00B
   calc_boundary_flux!       1.37k   15.4ms  0.76%  11.2μs     0.00B  0.00%    0.00B
   ~rhs!~                    1.37k   12.5ms  0.62%  9.14μs   1.24MiB  20.6%     948B
 analyze solution                3   85.4ms  4.22%  28.5ms   4.79MiB  79.4%  1.60MiB
 ───────────────────────────────────────────────────────────────────────────────────
```

with this PR (after compilation, Julia 1.7.0-beta4, `julia --check-bounds=no --threads=1`).

For reference, our standard DGSEM on curvilinear meshes yields
```julia
julia> trixi_include("examples/structured_2d_dgsem/elixir_euler_rayleigh_taylor_instability.jl",
           tspan=(0.0, 0.5), num_elements_per_dimension=16)
[...]
────────────────────────────────────────────────────────────────────────────────────────────────────
Trixi simulation finished.  Final time: 0.5  Time steps: 150 (accepted), 152 (total)
────────────────────────────────────────────────────────────────────────────────────────────────────

 ─────────────────────────────────────────────────────────────────────────────
           Trixi.jl                   Time                   Allocations      
                              ──────────────────────   ───────────────────────
       Tot / % measured:           1.84s / 92.4%           3.26MiB / 37.5%    

 Section              ncalls     time   %tot     avg     alloc   %tot      avg
 ─────────────────────────────────────────────────────────────────────────────
 rhs!                  1.37k    1.65s  97.3%  1.20ms    498KiB  39.8%     372B
   volume integral     1.37k    1.23s  72.3%   894μs     0.00B  0.00%    0.00B
   interface flux      1.37k    233ms  13.7%   170μs     0.00B  0.00%    0.00B
   boundary flux       1.37k   68.3ms  4.03%  49.8μs     0.00B  0.00%    0.00B
   surface integral    1.37k   53.0ms  3.13%  38.7μs     0.00B  0.00%    0.00B
   source terms        1.37k   26.8ms  1.58%  19.6μs    493KiB  39.4%     368B
   Jacobian            1.37k   18.6ms  1.10%  13.6μs     0.00B  0.00%    0.00B
   reset ∂u/∂t         1.37k   15.1ms  0.89%  11.0μs     0.00B  0.00%    0.00B
   ~rhs!~              1.37k   9.18ms  0.54%  6.70μs   5.14KiB  0.41%    3.84B
 analyze solution          3   46.3ms  2.73%  15.4ms    754KiB  60.2%   251KiB
 ─────────────────────────────────────────────────────────────────────────────
```

This changes the look and feel only very mildly - I tried to be minimally invasive. The basic idea is to re-order some multiplications using associativity to group things such that we can make more aggressive use of sparsity (aka sum factorization).