@jlchan 2021-09-13
If there is no difference for dense derivative operators, do you see any downside to making this the default for all specializations of calc_volume_integral! and removing the old version?

We can also do this in a separate PR if we want to merge this one faster.