@ranocha 2021-09-14
> @ranocha any update on this PR? The performance improvement is nice and the implementation is elegant. If you're busy at the moment, I can work on this too.

Sorry, family business yesterday. I plan to finish this today