@ranocha 2021-09-14
Thus, I would propose to merge this PR, benchmark possible versions for other volume integrals and make another PR if useful.