@jlchan 2021-09-14
> > If there is no difference for dense derivative operators, do you see any downside to making this the default for all specializations of `calc_volume_integral!` and removing the old version?
> 
> Do you mean adapting the other `DGMulti` volume integrals to use numerical fluxes with arbitrary coordinate direction (`normal_direction::AbstractVector`)? I'm not sure about that. It definitely pays off if you can use sparsity more efficiently. Otherwise, the numerical fluxes using Cartesian `direction`s are slightly more efficient, e.g.
> 
> ```julia
> julia> using BenchmarkTools, Trixi
> 
> julia> equations = CompressibleEulerEquations2D(1.4)
> ┌──────────────────────────────────────────────────────────────────────────────────────────────────┐
> │ CompressibleEulerEquations2D                                                                     │
> │ ════════════════════════════                                                                     │
> │ #variables: ………………………………………………… 4                                                                │
> │ │ variable 1: …………………………………………… rho                                                              │
> │ │ variable 2: …………………………………………… rho_v1                                                           │
> │ │ variable 3: …………………………………………… rho_v2                                                           │
> │ │ variable 4: …………………………………………… rho_e                                                            │
> └──────────────────────────────────────────────────────────────────────────────────────────────────┘
> 
> julia> normal_direction = SVector(1.0, 0.0)
> 2-element SVector{2, Float64} with indices SOneTo(2):
>  1.0
>  0.0
> 
> julia> u_ll = SVector(1.0, 0.5, -0.7, 1.0); u_rr = SVector(1.5, -0.2, 0.1, 5.0)
> 4-element SVector{4, Float64} with indices SOneTo(4):
>   1.5
>  -0.2
>   0.1
>   5.0
> 
> julia> @benchmark flux_ranocha($(Ref(u_ll))[], $(Ref(u_rr))[], $(Ref(1))[], $equations)
> BenchmarkTools.Trial: 10000 samples with 996 evaluations.
>  Range (min … max):  22.054 ns … 93.213 ns  ┊ GC (min … max): 0.00% … 0.00%
>  Time  (median):     22.307 ns              ┊ GC (median):    0.00%
>  Time  (mean ± σ):   23.263 ns ±  4.260 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%
> 
>   █▆▁▁▁                                                       ▁
>   █████▄▄▅▅▄▄▁▅▄▃▃▁▅▄▄▄▃▃▄▃▄▄▃▁███▇▆███▅▅▅▅▄▃▃▁▅▄▁▁▁▃▃▁▅▃▃▃▅▄ █
>   22.1 ns      Histogram: log(frequency) by time      48.4 ns <
> 
>  Memory estimate: 0 bytes, allocs estimate: 0.
> 
> julia> @benchmark flux_ranocha($(Ref(u_ll))[], $(Ref(u_rr))[], $(Ref(normal_direction))[], $equations)
> BenchmarkTools.Trial: 10000 samples with 996 evaluations.
>  Range (min … max):  24.045 ns … 71.146 ns  ┊ GC (min … max): 0.00% … 0.00%
>  Time  (median):     24.485 ns              ┊ GC (median):    0.00%
>  Time  (mean ± σ):   25.366 ns ±  4.105 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%
> 
>   ▆█▃                                 ▁                       ▁
>   █████▅▄▄▁▃▃▄▃▅▃▁▁▁▃▁▃▃▁▃▄▄▁▄▁▃▄▆▆▅▃▇█▇█▆▁▇█▇█▅▆▆▅▅▄▅▃▃▃▄▃▁▅ █
>   24 ns        Histogram: log(frequency) by time      45.8 ns <
> 
>  Memory estimate: 0 bytes, allocs estimate: 0.
> ```

Sounds good. It should be straightforward to extend the sum-factorization approach to all `DGMulti` solvers. Once this PR is merged I'll open a new PR to extend this and provide benchmarks.